// shared/src/types/game.ts
export type GameNames =
  | "Simon Says"
  | "connect Four"
  | "Rock Paper Scissors"
  | "Tic Tac Toe";

export const gameNames: GameNames[] = [
  "Simon Says",
  "connect Four",
  "Rock Paper Scissors",
  "Tic Tac Toe",
];
// players

// this is just to have interfaces comply with the same structure
export type Player = {
  id: string;
};

export interface CFplayer extends Player {
  choice: ConnectChoices;
}
export interface TTCPlayer extends Player {
  choice?: TTCOptions;
}
export interface RPSPlayer extends Player {
  choice: RPSOptions | null;
}
export interface SMSPlayer extends Player {}

// options
export type SMSColorType = "red" | "blue" | "green" | "yellow";
export const RPSOptionsValues = ["rock", "paper", "scissors"] as const;
export type RPSOptions = (typeof RPSOptionsValues)[number];
// ...
export type RPSRound = {
  winner: RPSPlayer;
  loser: RPSPlayer;
  isTie: boolean;
  moves: MoveChoice<RPSMove>[];
};

export type SMSRound = {
  sequence: SMSMove[];
};

export type Rounds = {
  count: number;
  rounds: RPSRound[];
  wins: {
    [key: string]: number;
    ties: number;
  };
};
// move

export type CFMove = MoveChoice<{
  color: ConnectChoices;
  coords: Coords;
}>;
export type RPSMove = MoveChoice<{
  choice: RPSOptions;
}>;
export type TTCMove = MoveChoice<{
  coords: Coords;
  choice: TTCOptions;
}>;
export type SMSMove = MoveChoice<{
  color: SMSColorType;
}>;
export type MoveChoice<T> = { id: string } & T;
// ...
export interface GameComponentProps {
  gameId: string;
  gameName: GameNames;
}